Parse refactoring & Config optimization
